Remove unused code and consolidate helpers
