+++ src/components/Tools/ConflictMediation/NeedsAssessment.js
@@ -52,91 +52,6 @@ const EMOTIONS = [
 
 // Styles are defined in NeedsAssessment.module.scss
 
-// Needs Content Component
-const NeedsContent = memo(({ isFullscreen }) => {
-	const [selectedLevel, setSelectedLevel] = useState(null);
-	const [selectedEmotion, setSelectedEmotion] = useState("");
-	const [notes, setNotes] = useState("");
-
-	const handleSubmit = (e) => {
-		e.preventDefault();
-		if (selectedLevel && selectedEmotion) {
-			alert(`Check-in saved!\nLevel: ${selectedLevel.name}\nEmotion: ${selectedEmotion}\nNotes: ${notes}`);
-			setSelectedLevel(null);
-			setSelectedEmotion("");
-			setNotes("");
-		}
-	};
-
-  return (
-    <div className={styles.container}>
-      <h1 className={styles.title}>Daily Maslow Check-In</h1>
-
-      <div className={styles.contentSection}>
-        <h2 className={styles.sectionTitle}>Select Level</h2>
-        <div className={styles.levelsList}>
-          {MASLOW_LEVELS.map((level) => (
-            <button
-              key={level.name}
-              type="button"
-              className={`${styles.levelItem} ${selectedLevel === level ? styles.active : ''}`}
-              onClick={() => setSelectedLevel(level)}
-            >
-              <div className={styles.levelHeader}>
-                <span>{level.emoji}</span>
-                <span className={styles.levelName} style={{ color: level.color }}>{level.name}</span>
-              </div>
-              <div className={styles.levelDescription}>{level.description}</div>
-            </button>
-          ))}
-        </div>
-      </div>
-
-                        {selectedLevel && (
-                                <div className={styles.contentSection}>
-                                        <h2 className={styles.sectionTitle}>How do you feel?</h2>
-                                        <div className={styles.emotionsGrid}>
-                                                {EMOTIONS.map((emotion) => (
-                                                        <button
-                                                                key={emotion}
-                                                                type="button"
-                                                                className={`${styles.emotionButton} ${selectedEmotion === emotion ? styles.active : ''}`}
-                                                                onClick={() => setSelectedEmotion(emotion)}
-                                                        >
-                                                                {emotion}
-                                                        </button>
-                                                ))}
-                                        </div>
-                                </div>
-                        )}
-
-                        {selectedLevel && selectedEmotion && (
-                                <div className={styles.contentSection}>
-                                        <h2 className={styles.sectionTitle}>Additional Notes</h2>
-                                        <textarea
-                                                className={styles.notesTextarea}
-                                                value={notes}
-                                                onChange={(e) => setNotes(e.target.value)}
-                                                placeholder="Any thoughts or reflections..."
-                                        />
-                                </div>
-                        )}
-
-                        <div className={styles.contentSection}>
-                                <button
-                                        className={styles.submitButton}
-                                        onClick={handleSubmit}
-                                        disabled={!selectedLevel || !selectedEmotion}
-                                >
-                                        Save Check-In
-                                </button>
-                        </div>
-                </div>
-        );
-});
-
-NeedsContent.displayName = "NeedsContent";
-
 // Main component
 const NeedsAssessment = ({ onNeedsSelected }) => {
 	const [selectedNeeds, setSelectedNeeds] = useState([]);
