+++ src/components/Tools/shared/hooks/index.js
@@ -2,6 +2,6 @@ export * from '../../../effects/Matrix/AuthContext';
 export { 
   useVisibilityObserver,
   useScreenOrientation,
-  useMobileDetection,
-  useKeyboardNavigation 
-} from '../hooks'; 
+  useKeyboardNavigation
+} from '../hooks';
+
